HT7

+++ src/components/Article/index.js
@@ -9,22 +9,26 @@ import Loader from '../Loader'
 
 class Article extends Component {
     static propTypes = {
+        id: PropTypes.string,
+        isOpen: PropTypes.bool,
+        toggleOpen: PropTypes.func,
+        handleDelete: PropTypes.func,
+        //from connect
         article: PropTypes.shape({
             title: PropTypes.string.isRequired,
             text: PropTypes.string,
             comments: PropTypes.array
-        }).isRequired,
-        isOpen: PropTypes.bool,
-        toggleOpen: PropTypes.func,
-        handleDelete: PropTypes.func.isRequired
+        })
     }
 
-    componentWillReceiveProps({isOpen, article, loadArticle}) {
-        if (!this.props.isOpen && isOpen) loadArticle(article.id)
+    componentDidMount() {
+        const {isOpen, id, loadArticle} = this.props
+        if (isOpen) loadArticle(id)
     }
 
     render() {
         const { article, toggleOpen } = this.props
+        if (!article) return null
         return (
             <div>
                 <h3 onClick = {toggleOpen}>{article.title}</h3>
@@ -42,6 +46,7 @@ class Article extends Component {
 
     deleteArticle = () => {
         const {handleDelete, article} = this.props
+        if (!handleDelete || !article) return
         handleDelete(article.id)
     }
 
@@ -52,10 +57,12 @@ class Article extends Component {
         return (
             <div>
                 <p>{article.text}</p>
-                <CommentList article = {article} />
+                <CommentList article = {article} ref = "commentList" />
             </div>
         )
     }
 }
 
-export default connect(null, { loadArticle })(Article)
+export default connect((state, {id}) => ({
+    article: state.articles.getIn(['entities', id])
+}), { loadArticle })(Article)
